more balance

+++ Assets/Scripts/Abilities/AbilityEffects/EffectScripts/HealEffect.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class AbilityHealEffect : AbilityEffectBase {
 
     public override IEnumerator ApplyToCharacter(AbilityContext context) {
-        int heal = (int)((context.user.Character.Stats.Magic * (2 + context.user.Character.Level)/(3 + 1) +100) / (100 * context.ability.Power));
+        int heal = (int)(((float)context.user.Character.Stats.Magic * (2.0f + (float)context.user.Character.Level)/(3.0f + 1.0f) + 100.0f) / (100.0f * (float)context.ability.Power));
         context.target.Character.IncreaseHP(heal);
         GameObject damageTextObject = context.target.CurrentModelInstance.transform.GetChild(0).gameObject;
             damageTextObject.SetActive(true);
